feat(host): allow for a configurable host
replace the environment that uses static url by a string in which
the users can put any host they want.

BREAKING CHANGE:
This change is breaking because it changes Environments enum into a string.
I've kept the constant around to minimize the impact but if users are using
the Environments enum, they'll have to fix their build.

I believe this is a minor change but breaking nonetheless so it will come
with a major version bump.