docs(token): configuration with an app token
Previously, there was no app token. At that time, static token
meant:

1. fetching a token using the /oauth/token API
2. copy/paste it in the code

That's why it was deemed insecure. With app tokens, a user can
generate a valid token with restricted access. It's for a
different application but it works fine and it's not more/less
secure than using a key/secret.

[skip ci]